[JS-24] feat (kits): implement toaster use google calendar api
