+++ .gitignore
@@ -18,6 +18,8 @@
 .env.test.local
 .env.production.local
 
+apiGoogleconfig.json
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
