+++ package.json
@@ -15,6 +15,8 @@
     "node-sass": "^6.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-google-calendar-api": "^1.5.2",
+    "react-google-login": "^5.2.2",
     "react-hook-form": "^7.13.0",
     "react-icons": "^4.2.0",
     "react-router-dom": "^5.2.1",
