--- src/components/Kits/EncontrosData.js
@@ -1,42 +1,0 @@
-
-
-export const EncontrosData = [
-    {
-        title: 'Encontro para a entrega',
-        startTime: 'December 20, 2021 21:00:00 UTC',
-        
-        hour: '20:00',
-        cName: 'hexa',
-        dia: '20/12/2021'
-    },
-
-    {
-        title: ' Meet de natal',
-        startTime: 'December 25, 2021 17:00:00 UTC',
-        hour: '17:00',
-        cName: 'hexa',
-        dia: '25/12/2021'
-    },
-    {
-        title: ' Daily c marquinhos',
-        startTime: 'December 23, 2021 10:00:00 UTC',
-        hour: '10:00',
-        cName: 'hexa',
-        dia: '23/12/2021'
-    },
-    {
-        title: ' Scrum nov 2',
-        startTime: 'October 8, 2021 10:00:00 UTC',
-        hour: '10:00',
-        cName: 'hexa',
-        dia: '08/10/2021'
-    },
-    {
-        title: ' Scrum Nov 1',
-        startTime: 'December 2, 2021 10:00:00 UTC',
-        hour: '10:00',
-        cName: 'hexa',
-        click: 'calendarEvent',
-        dia: '02/12/2021'
-    }
-]
