+++ src/components/Kits/encontros.scss
@@ -42,6 +42,7 @@
     display: flex;
     flex-direction: row;
     width: 100px;
+    cursor: pointer;
 }
 
 .button-menu.active {
