--- src/components/pontosDeEncontro/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-class Button extends React.Component {
-    render () {
-        return (
-            <button>Nome do button</button>
-        )
-    }
-}
-
-export default Button;
-
-
-
